% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_check.R
\name{conda_check}
\alias{conda_check}
\title{Check if Conda is Installed}
\usage{
conda_check()
}
\value{
A logical value: \code{TRUE} if Conda is installed, \code{FALSE} otherwise.
}
\description{
This function checks if Conda is installed on the system by looking for the \code{conda}
command in the system's PATH. It is used internally by the \code{conda_whats_installed} function.
}
\details{
The function will return \code{TRUE} if Conda is found in the system's PATH, meaning that Conda
is installed and accessible. If Conda is not found, it returns \code{FALSE}, indicating that
Conda is not installed.
}
