% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_dl.R
\name{conda_dl}
\alias{conda_dl}
\title{Download and Install Miniconda}
\usage{
conda_dl()
}
\value{
A character vector with the status of each step in the process, including
any error messages encountered during the download, installation, or path setup.
}
\description{
This function  downloads and installs Miniconda, note that it is hard coded to only download the Intel MacOSX installation package.
}
\details{
The function first checks if Conda is already installed. If it is, it
will return a message indicating that Conda is already installed and set up. If Conda
is not installed, it will proceed to download and install Miniconda. The installation
process is done via a bash script, and any errors during the download or installation
will be captured and reported. After the installation, the function also ensures that
the Conda path is correctly set in the system's environment.

The function will return a status report as a character vector, which will indicate
the success or failure of each step in the process.
}
